requirements.txt
fastapi
uvicorn
anthropic
python-dotenv
requests
beautifulsoup4
schedule
